MINIFICPP-1096 review fixes

+++ extensions/http-curl/client/HTTPStream.cpp
@@ -59,7 +59,7 @@ int HttpStream::writeData(std::vector<uint8_t> &buf, int buflen) {
     throw minifi::Exception{ExceptionType::GENERAL_EXCEPTION, "negative buflen"};
   }
 
-  if (buf.size() < buflen) {
+  if (buf.size() < static_cast<size_t>(buflen)) {
     return -1;
   }
   return writeData(buf.data(), buflen);
